Update README.md
KION India Pvt. Ltd. is a reliable name for its customer when it comes to designing, manufacturing, distribution of services, and supporting Palletized Material Handling Equipment under the brand name OM for the Indian market. The brand looks forward to sustainably catering to the needs of an ever modernizing Indian manufacturing industry.